Reorganize file structure.
The include files are now placed in a subdirectory, so that they end up
in a sub-directory of the target directory. This helps to avoid name
collisions.

The generated library and DBD file have also been renamed so that they
both have the same name and match the scheme of the ChimeraTK Control
System Adapter core library.