Call str method through reference to ostringstream.
The str() method only exists in the std::ostringstream class, not in the
std::basic_ostream class, so calling on the result of the << operator
(which returns a reference to std::basic_ostream) does not work.
On macOS, this code compiled (presumably because the standard library
overloads the << operator in std::ostringstream so that it returns a
reference of the same type), but it did not compile on Linux (the GNU
standard library does not seem to implement it this way).