Add support for the ApplicationBase class.
The ApplicationBase class has been introduced to the ChimeraTK
control-system adapter so that application libraries can register a
point of entry without needing any code that depends on both the
application and the control-system in use.

This patch introduces the chimeraTKConfigureApplication IOC shell
command so that such applications can be used with EPICS without
having to write any C or C++ code.