Add ChimeraTK libraries to _SYS_LIBS.
The are already includes through USR_LDFLAGS, but those are added early
in the command line and thus the resulting device support library is not
linked to the ChimeraTK libraries when compiling on Linux. This has the
consequence that an IOC linking to the device support will not compile
unless it explicitly links to the ChimeraTK libraries as well.