Use ReadAnyGroup instead of private APIs.
With the latest changes to the ReadAnyGroup, we can actually use its API
in order to wait for update notifications.

Please note that this change also changes the behavior slightly:
Notifications are now only received until we get a notification for a PV
for which the last notification has not been completely processed yet.
If that happens, we wait until processing has finished for this PV
before receiving the next notification.