Add support for bidirectional process variables.
With the changes from this changeset, output records will now be
notified when a bidirectional process variable is changed on the
application / device side. This means that the user has now a chance to
see that a value that was set earlier is not active any longer.

As a side effect, the changes made as part of this changeset also bring
two other improvements:

First, an input record will now receive an initial update when switching
it to I/O Intr mode. Before, it would only receive an update when the
application or device actually sent an update notification, meaning that
the record might potentially contain an outdated value for a long time.

Second, when output records are initialized, the INVALID alarm on them
is now removed in addition to setting the UDF field to zero.