Avoid use of uninitialized lastValue (fixes #2).
This patch also refactors the way how initialValue() works. Instead of
keeping track of whether the ProcessArray still contains a valid value
or the lastValue should be used, we move the ProcessArray's value to
lastValue during initialization, so that lastValue always contains a
valid value. This way, we can remove the whole logic around
initialValueAvailable, simplifying the code considerably.