Make code compatible with EPICS Base 7.0.6.
This necessitates two changes: First, EPICS Base’s dbLink.h includes
other headers that contain C++ code in recent releases of EPICS Base, so
it must no longer be included from with an “extern "C"” block.

Second, the tools for building DBD files now check dependencies, so we
have to use “DBDCAT” instead of “DBD”, because we are only building
files that shall later be included in an IOC, so we do not want to
include these dependencies.